+++ crates/kornia-lie/src/se3.rs
@@ -496,8 +496,8 @@ mod tests {
         let mut adj_product = [[0.0f32; 6]; 6];
         for i in 0..6 {
             for j in 0..6 {
-                for k in 0..6 {
-                    adj_product[i][j] += adj_1[i][k] * adj_2[k][j];
+                for (k, adj_2k) in adj_2.iter().enumerate() {
+                    adj_product[i][j] += adj_1[i][k] * adj_2k[j];
                 }
             }
         }
