+++ crates/kornia-nn/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "kornia-nn"
+description = "Neural network operators in Rust"
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+include = { workspace = true }
+license = { workspace = true }
+readme = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+version = { workspace = true }
+
+[dependencies]
+matrixmultiply = "0.3"
+
+[dev-dependencies]
+approx = { workspace = true }
+criterion = { workspace = true }
+rand = { workspace = true }
+wide = "0.7.32"
+gemm = "=0.18.2"
+candle-nn = { version = "0.9.1", features = ["mkl"] }
+candle-core = { version = "0.9.1", features = ["mkl"] }
+
+[[bench]]
+name = "bench_linear"
+harness = false
