+++ crates/kornia-nn/src/lib.rs
@@ -1,0 +1,5 @@
+#![deny(missing_docs)]
+//! # Kornia Neural Network operations
+
+/// Linear layer operations
+pub mod linear;
