+++ crates/kornia-tensor-ops/src/lib.rs
@@ -1,13 +1,10 @@
 #![deny(missing_docs)]
 #![doc = env!("CARGO_PKG_DESCRIPTION")]
 
-/// Deep neural network kernels.
-pub mod dnn;
-
-/// Error types for the core-ops module.
+/// Error types for the tensor-ops module.
 pub mod error;
 
-/// module containing the kernels for the tensor operations.
+/// module containing the kernels for the tensor operations. (e.g. `kornia_tensor::ops::add`)
 pub mod kernels;
 
 /// module containing ops implementations.
